requirements.txt
dnspython==2.2.0
fastapi==0.74.0
sslyze==5.0.2